requirements.txt
# LexSimple Dependencies
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0

# LLM
google-genai>=1.0.0

# Document parsing
pdfplumber>=0.10.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
